rename analytics-mixin to just analytics, getting rid of redundancy.
