+++ app/mixins/analytics.js
