+++ app/routes/index.js
@@ -1,9 +1,9 @@
 import Ember from 'ember';
 
 import ResetScrollMixin from '../mixins/reset-scroll';
-import AnalyticsMixin from '../mixins/analytics-mixin';
+import Analytics from '../mixins/analytics';
 
-export default Ember.Route.extend(AnalyticsMixin, ResetScrollMixin, {
+export default Ember.Route.extend(Analytics, ResetScrollMixin, {
     model() {
         return this.store.query('taxonomy', { filter: { parents: 'null' }, page: { size: 20 } });
     },
